+++ cli/src/core/ast/deletions.rs
@@ -1,8 +1,8 @@
+pub(crate) mod delete_from_client_sdk;
 pub(crate) mod delete_from_index_ts;
 pub(crate) mod delete_from_registrations_ts;
 pub(crate) mod delete_from_sdk_ts;
 pub(crate) mod delete_from_seed_data_ts;
 pub(crate) mod delete_from_server_ts;
 pub(crate) mod delete_from_test_utils;
-pub(crate) mod delete_from_universal_sdk;
 pub(crate) mod delete_import_statement;